pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["alanwoodenclouds <alan.woodenclouds@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
django-cors-headers = "^4.4.0"
djangorestframework = "^3.15.2"
cryptography = "^43.0.0"
djangorestframework-simplejwt = "^5.3.1"
python-decouple = "^3.8"
pillow = "^10.4.0"
dj-database-url = "^2.2.0"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
